Implements base interaction UI
Adds a UGUI element for player interaction with bases.

Introduces a new UXML file for the interaction UI, including a button to enter the base. Includes a new script "BaseUIController" for managing the UI's visibility and interaction. It uses DOTween for animations to smoothly show and hide the UI element when the player clicks on the base.

Also adds a "UIBillboard" script to maintain the UI's orientation towards the camera.